(PATCH) - Adjust the root styles
This commit represents a fix to lessons "The Root Route", "Using Tailwind Part 1" and "Using Tailwind Part 3". We shouldn't have put the flex class on the body element because it makes Remix's default error UI look weird.

Instead, we should put the h-screen class on the body and the flex and h-screen classes on a div wrapping the App component.