Fix loader caching
Now that we're using React Router v7's Single Fetch feature, we have to cache loader data differently.

With Single Fetch, data for all matched loaders is combined into a single response. Along with this, headers for data requests are now handled by the `headers` export, which previously only handled document request headers.

These changes mean that we have to move cache-control headers for loader data responses into the `headers` export.

For caching responses that use etags, we now have to validate the etag and send an empty 304 response in the `entry.server.tsx` file in the `handleDataRequest` export, rather than in directly in the loader.